feat: implement required functionality

+++ domain/src/main/java/com/itexus/domain/model/Currency.kt
@@ -1,8 +1,6 @@
 package com.itexus.domain.model
 
-import java.math.BigDecimal
-
 data class Currency(
     val name: String,
-    val value: BigDecimal
+    val value: Double
 )
