+++ domain/src/main/java/com/itexus/domain/model/Rate.kt
@@ -1,9 +1,7 @@
 package com.itexus.domain.model
 
-import java.math.BigDecimal
-
 data class Rate(
     val name: String,
     val base: String,
-    val value: BigDecimal
+    val value: Double,
 )
